Change:プロバイダーの定義ファイルを3つに分割
