+++ test/query_provider_test.dart
@@ -8,7 +8,9 @@ import 'package:mocktail/mocktail.dart';
 import 'package:yumemi_codecheck_repo_search/generated/l10n.dart';
 import 'package:yumemi_codecheck_repo_search/model/repo.dart';
 import 'package:yumemi_codecheck_repo_search/model/repo_search_result.dart';
-import 'package:yumemi_codecheck_repo_search/service.dart';
+import 'package:yumemi_codecheck_repo_search/provider/search_query_provider.dart';
+import 'package:yumemi_codecheck_repo_search/provider/search_result_provider.dart';
+import 'package:yumemi_codecheck_repo_search/provider/service_provider.dart';
 import 'package:yumemi_codecheck_repo_search/service/github_repo_service.dart';
 
 import 'mocks.dart';
